backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib==1.7.4
bcrypt==3.2.0
python-jose[cryptography]
pydantic[email]
python-multipart